+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning].
 
 - /
 
+## [1.1.0] - 2023-07-01
+
+- Add ResolverInterface. With this feature, now it's possible to add custom resolvers to the container.
+- Move arrays of definitions to the new helper class Bag.
+
 ## [1.0.0] - 2023-06-26
 
 - initial release
@@ -20,5 +25,6 @@ and this project adheres to [Semantic Versioning].
 
 <!-- Versions -->
 
-[unreleased]: https://github.com/iziphp/container/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/iziphp/container/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/iziphp/container/releases/tag/v1.1.0
 [1.0.0]: https://github.com/iziphp/container/releases/tag/v1.0.0